Changed package names from com.example.fakturapp to com.fakturapp
