+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,10 @@
         android:theme="@style/Theme.FakturApp"
         tools:targetApi="31">
         <activity
-            android:name=".EditProfile"
+            android:name="com.fakturapp.EditProfile"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.fakturapp.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
