--- app/src/main/java/com/example/fakturapp/SignIn.java
@@ -1,4 +1,0 @@
-package com.example.fakturapp;
-
-public class SignIn {
-}
