--- app/src/main/java/com/example/fakturapp/SignUp.java
@@ -1,4 +1,0 @@
-package com.example.fakturapp;
-
-public class SignUp {
-}
