+++ app/src/main/java/com/fakturapp/MainActivity.java
@@ -1,16 +1,16 @@
-package com.example.fakturapp;
+package com.fakturapp;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.provider.FontRequest;
 
 import android.os.Bundle;
 
-public class EditProfile extends AppCompatActivity {
+import com.example.fakturapp.R;
 
+public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_edit_profile);
+        setContentView(R.layout.activity_main);
     }
 }
