+++ app/src/main/java/com/fakturapp/SignIn.java
@@ -1,0 +1,4 @@
+package com.fakturapp;
+
+public class SignIn {
+}
