+++ app/src/main/res/layout/activity_add_client.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditProfile">
+    tools:context="com.fakturapp.EditProfile">
 
     <TextView
         android:id="@+id/textView2"
@@ -110,13 +110,15 @@
         android:layout_height="45dp"
         android:layout_marginTop="136dp"
         android:ems="10"
+        android:fontFamily="@font/roboto"
         android:inputType="textPersonName"
         android:text="Jeżeli nie jest przedsiębiorcą, pomiń ten krok"
         android:textColor="#979393"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        tools:text="Jeżeli nie jest przedsiębiorcą, pomiń ten krok" />
 
     <TextView
         android:id="@+id/textView4"
