+++ app/src/main/res/layout/activity_edit_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditProfile">
+    tools:context="com.fakturapp.EditProfile">
 
     <TextView
         android:id="@+id/textView2"
