+++ app/src/test/java/com/fakturapp/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.fakturapp;
+package com.fakturapp;
 
 import org.junit.Test;
